ml/requirements.txt
joblib==1.4.2
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
protobuf==4.21.12
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
threadpoolctl==3.5.0
tzdata==2024.1
